requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
jinja2
passlib
email-validator
qrcode
pillow
razorpay
python-dotenv